Updated external sources to current versions modified   ETSLayout/ETSLayout.cpp modified   ETSLayout/ETSLayout.h added      gridctrl/DropDownButton.cpp added      gridctrl/DropDownButton.h modified   gridctrl/GridCtrl.cpp modified   gridctrl/GridCtrl.h added      gridctrl/ListBoxDrop.cpp added      gridctrl/ListBoxDrop.h added      gridctrl/ModGridCtrl.cpp added      gridctrl/ModGridCtrl.h added      gridctrl/ModGridCtrlEx.cpp added      gridctrl/ModGridCtrlEx.h modified   gridctrl/TitleTip.cpp added      sizecbar/scbarcfvs7.cpp added      sizecbar/scbarcfvs7.h added      sizecbar/sdbarcfvs7.cpp added      sizecbar/sdbarcfvs7.h modified   sizecbar/sizecbar.cpp modified   VisualStylesXP/VisualStylesXP.h
git-svn-id: svn://136.177.114.72/svn_GW/WPhast/trunk@208 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ gridctrl/GridCtrl.cpp
@@ -1834,7 +1834,7 @@ COleDataSource* CGridCtrl::CopyTextFromGrid()
     char c = '\0';
     sf.Write(&c, 1);
 
-    DWORD dwLen = sf.GetLength();
+    DWORD dwLen = (DWORD)sf.GetLength();
     HGLOBAL hMem = sf.Detach();
     if (!hMem) 
         return NULL;
